rest_framework_project/first_project/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-filter==22.1
django-versatileimagefield==2.2
djangorestframework==3.13.1
drf-flex-fields==1.0.0
install==1.3.5
Markdown==3.4.1
Pillow==9.2.0
python-magic==0.4.27
pytz==2022.2.1
sqlparse==0.4.2
